provide hovers + enhance heuristics for definitions
If `basicCodeIntel.hover` is enabled in user settings, this extension will provide hovers consisting of:

- the best-guess definition line's code, cleaned up a bit (e.g., removing trailing `:`s)
- the docstring before or after the definition line

It only shows a hover if the definition location can be found using symbols. If no results are found using symbols, it does not fall back to text search for generating hover contents because that is too inaccurate.

The `basicCodeIntel.hover` setting is off by default.