feat: enable defs/refs/hovers by default
Previously, all functionality was disabled by default (except for command palette actions). This meant that users needed to explicitly run "Enable fuzzy defs/refs" and "Enable fuzzy hovers" in the command palette after enabling this extension. It is easy for users to not know this was necessary and therefore not understand how to use this extension. Now, it enables them by default.

fix https://github.com/sourcegraph/sourcegraph/issues/1779