Enable search-based intelligence for Objective C.
Since Objective C shares header file extensions with C and C++, this commit
treats Objective C as a dialect of C++ for the purpose of code intelligence.
In addition to adding the .m file extension, it extends the import filter to
allow #import and @import directives as well as #include.

For native C and C++ code, these extensions will have no material effect.