Do not emit event on empty definition result set (#334)
It turns out we don't show a hover for empty location sets, but we are still emitting the event when we get back an empty location set from the language server. This happens particularly when someone mouses over a position in the that looks like a symbol from the UI's point of view, but isn't really anything from the lang server's point of view (maybe an identifier-like token in a string or comment).

This will stop sending non-interesting results on lspDefinition. LSP references and hovers are already guarded, as are all LSIF results. This was the odd-one-out.