search: Set explicit count:500 parameter in referencesQuery
So far we relied on an implicit count being assumed. On the GraphQL API
that is 30. However, we had a bug that sometimes made us return more
than 30 results. When this bug was fixed in https://github.com/sourcegraph/sourcegraph/pull/29281,
the regression tests for find references failed, since we were now
cutting properly, and one of the expected references was then missing.

In general, @efritz commented that we assumed a `count:all`. That was
never the case, and doing so could be prohibitively expensive on
sourcegraph.com. So we instead ask for the display limit of 500 which is
already enforced.